made continue count visible on continue button increased continue count to 2

+++ Assets/Scripts/GameOver/GameOverSettings.cs
@@ -6,7 +6,7 @@ namespace GameOver
     [Serializable]
     public class GameOverSettings
     {
-        [field: SerializeField] public float MaxContinueCount { get; private set; }
+        [field: SerializeField] public int MaxContinueCount { get; private set; }
         [field: SerializeField] public float ClearSmallTilesBiggestTileMultiplier { get; private set; }
     }
 }
